Refactor pruner scope tab into focused subcomponents
Break the large pruner scope page into controller and focused UI sections for default/provider views while preserving behavior and tests.

Closes #134
